omfg fix pay
